Lazy loading modules

+++ slt/src/app/pages/authenticated/admin/usermanagement/usermanagement.component.ts
@@ -9,17 +9,17 @@ import { ToastService } from 'src/app/shared/services/toast.service';
   styleUrls: ['./usermanagement.component.scss'],
 })
 export class UserManagementComponent implements OnInit {
-  public allUsers: UserAccount[];
+  public allUsers: UserAccount[] = []
   public isModalVisible = false;
-  public userName: string;
+  public userName: string = '';
   public displayedColumns: string[] = ['id', 'name', 'email', 'delete'];
 
-  public selectedUser: UserAccount;
+  public selectedUser: UserAccount | undefined;
 
   constructor(
     private adminService: AdminService,
     private toastService: ToastService
-  ) {}
+  ) { }
 
   ngOnInit() {
     this.getAllUsers();
@@ -46,18 +46,20 @@ export class UserManagementComponent implements OnInit {
   }
 
   public deleteUser() {
-    this.adminService.deleteUser(this.selectedUser).subscribe(
-      () => {
-        this.toastService.setMessage('User account successfully deleted');
-        this.getAllUsers();
-      },
-      () => {
-        this.toastService.setMessage('User account could not be deleted');
-      },
-      () => {
-        this.selectedUser = undefined;
-        this.closeModal();
-      }
-    );
+    if (this.selectedUser) {
+      this.adminService.deleteUser(this.selectedUser).subscribe(
+        () => {
+          this.toastService.setMessage('User account successfully deleted');
+          this.getAllUsers();
+        },
+        () => {
+          this.toastService.setMessage('User account could not be deleted');
+        },
+        () => {
+          this.selectedUser = undefined;
+          this.closeModal();
+        }
+      );
+    }
   }
 }
